requirements.txt
flask
pymongo
python-dotenv
flask-cors
flask-jwt-extended
bcrypt
